requirements.txt
alembic==1.13.1
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.53
Werkzeug==2.3.7
python-jose==3.3.0
PyJWT==2.8.0
